+++ transformer.py
@@ -5,17 +5,18 @@ from MultiHeadAttention import MultiHeadAttention
 class Transformer(torch.nn.Module):
     def __init__(self):
         super().__init__()
-        self.normalization_layer = torch.nn.LayerNorm(EMBEDDING_DIM, dtype=torch.float64)
-        self.normalization_layer_2 = torch.nn.LayerNorm(EMBEDDING_DIM, dtype=torch.float64)
+        self.normalization_layer = torch.nn.LayerNorm(EMBEDDING_DIM, dtype=torch.float32)
+        self.normalization_layer_2 = torch.nn.LayerNorm(EMBEDDING_DIM, dtype=torch.float32)
 
         self.attention_layer = MultiHeadAttention()
         
         # only turn on for gpu
-        self.expand = torch.nn.Linear(EMBEDDING_DIM, (SCALE * EMBEDDING_DIM), dtype=torch.float64)
+        self.expand = torch.nn.Linear(EMBEDDING_DIM, (SCALE * EMBEDDING_DIM), dtype=torch.float32)
         self.relu_layer = torch.nn.ReLU()
-        self.contract = torch.nn.Linear((SCALE * EMBEDDING_DIM), (EMBEDDING_DIM), dtype=torch.float64)
+        self.contract = torch.nn.Linear((SCALE * EMBEDDING_DIM), (EMBEDDING_DIM), dtype=torch.float32)
 
     def forward(self, embeddings, B_matrix):
+
         normalized_embeddings = self.normalization_layer(embeddings)
         attention_embeddings = self.attention_layer(normalized_embeddings, B_matrix)
         skip_embed = embeddings + attention_embeddings
